+++ rohrpost/main.py
@@ -1,33 +1,30 @@
 import json
 from functools import partial
 
-from . import handlers  # noqa
+from channels.generic.websocket import WebsocketConsumer
+
+from . import handlers
 from .message import send_error
 from .registry import HANDLERS
+
+assert handlers  # silence qa
 
 REQUIRED_FIELDS = ["type", "id"]
 
 
-try:
-    DECODE_ERRORS = (json.JSONDecodeError, TypeError)
-except AttributeError:
-    # Python 3.4 raises a ValueError instead of json.JSONDecodeError
-    DECODE_ERRORS = (ValueError, TypeError)
-
-
-def handle_rohrpost_message(message):
+def handle_rohrpost_message(consumer: WebsocketConsumer, text_data: str) -> None:
     """
     Handling of a rohrpost message will validate the required format:
     A valid JSON object including at least an "id" and "type" field.
     It then hands off further handling to the registered handler (if any).
     """
-    _send_error = partial(send_error, message=message, message_id=None, handler=None)
-    if not message.content["text"]:
+    _send_error = partial(send_error, consumer=consumer, message_id=None, handler=None)
+    if not text_data:
         return _send_error(error="Received empty message.")
 
     try:
-        request = json.loads(message.content["text"])
-    except DECODE_ERRORS as e:
+        request = json.loads(text_data)
+    except (json.JSONDecodeError, TypeError) as e:
         return _send_error(
             error="Could not decode JSON message. Error: {}".format(str(e))
         )
@@ -39,12 +36,13 @@ def handle_rohrpost_message(message):
         if field not in request:
             return _send_error(error="Missing required field '{}'.".format(field))
 
-    if not request["type"] in HANDLERS:
+    request_type = request["type"]
+    if request_type not in HANDLERS:
         return send_error(
-            message=message,
+            consumer=consumer,
             message_id=request["id"],
-            handler=request["type"],
-            error="Unknown message type '{}'.".format(request["type"]),
+            handler=request_type,
+            error="Unknown message type '{}'.".format(request_type),
         )
 
-    HANDLERS[request["type"]](message, request)
+    HANDLERS[request_type](consumer=consumer, request=request)