+++ tests/test_message.py
@@ -3,26 +3,26 @@ import pytest
 from rohrpost.message import send_error, send_message, send_success
 
 
-def test_send_error(message):
+def test_send_error(consumer):
     send_error(
-        message=message, message_id=1, handler="some_handler", error="error_message"
+        consumer=consumer, message_id=1, handler="some_handler", error="error_message"
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert data[0]["id"] == 1
     assert data[0]["type"] == "some_handler"
     assert data[0]["error"] == "error_message"
 
 
-def test_send_error_with_additional_data(message):
+def test_send_error_with_additional_data(consumer):
     send_error(
-        message=message,
+        consumer=consumer,
         message_id=1,
         handler="some_handler",
         error="error_message",
         data={"some_field": "additional_info"},
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert data[0]["id"] == 1
     assert data[0]["type"] == "some_handler"
@@ -30,26 +30,29 @@ def test_send_error_with_additional_data(message):
     assert data[0]["data"]["some_field"] == "additional_info"
 
 
-def test_send_error_without_id(message):
+def test_send_error_without_id(consumer):
     send_error(
-        message=message, message_id=None, handler="some_handler", error="error_message"
+        consumer=consumer,
+        message_id=None,
+        handler="some_handler",
+        error="error_message",
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert "id" not in data[0]
     assert data[0]["type"] == "some_handler"
     assert data[0]["error"] == "error_message"
 
 
-def test_send_error_without_id_with_additional_data(message):
+def test_send_error_without_id_with_additional_data(consumer):
     send_error(
-        message=message,
+        consumer=consumer,
         message_id=None,
         handler="some_handler",
         error="error_message",
         data={"some_field": "additional_info"},
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert "id" not in data[0]
     assert data[0]["type"] == "some_handler"
@@ -57,15 +60,15 @@ def test_send_error_without_id_with_additional_data(message):
     assert data[0]["data"]["some_field"] == "additional_info"
 
 
-def test_send_error_without_id_and_type(message):
+def test_send_error_without_id_and_type(consumer):
     send_error(
-        message=message,
+        consumer=consumer,
         message_id=None,
         handler=None,
         error="error_message",
         data={"some_field": "additional_info"},
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert "id" not in data[0]
     assert "type" not in data[0]
@@ -73,31 +76,31 @@ def test_send_error_without_id_and_type(message):
     assert data[0]["data"]["some_field"] == "additional_info"
 
 
-def test_send_error_without_anything(message):
-    send_error(message=message, message_id=None, handler=None, error="error_message")
-    data = message.reply_channel.data
+def test_send_error_without_anything(consumer):
+    send_error(consumer=consumer, message_id=None, handler=None, error="error_message")
+    data = consumer.data
     assert len(data) == 1
     assert "id" not in data[0]
     assert "type" not in data[0]
     assert data[0]["error"] == "error_message"
 
 
-def test_send_success(message):
-    send_success(message=message, message_id=1, handler="some_handler")
-    data = message.reply_channel.data
+def test_send_success(consumer):
+    send_success(consumer=consumer, message_id=1, handler="some_handler")
+    data = consumer.data
     assert len(data) == 1
     assert data[0]["id"] == 1
     assert data[0]["type"] == "some_handler"
 
 
-def test_send_success_with_additional_data(message):
+def test_send_success_with_additional_data(consumer):
     send_success(
-        message=message,
+        consumer=consumer,
         message_id=1,
         handler="some_handler",
         data={"some_field": "additional_info"},
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert data[0]["id"] == 1
     assert data[0]["type"] == "some_handler"
@@ -105,16 +108,16 @@ def test_send_success_with_additional_data(message):
     assert "some_field" not in data[0]
 
 
-def test_send_success_without_id(message):
+def test_send_success_without_id(consumer):
     with pytest.raises(Exception) as exc:
-        send_success(message=message, message_id=None, handler="some_handler")
+        send_success(consumer=consumer, message_id=None, handler="some_handler")
     assert "ID and handler" in str(exc.value)
 
 
-def test_send_success_without_id_with_additional_data(message):
+def test_send_success_without_id_with_additional_data(consumer):
     with pytest.raises(Exception) as exc:
         send_success(
-            message=message,
+            consumer=consumer,
             message_id=None,
             handler="some_handler",
             data={"some_field": "additional_info"},
@@ -122,10 +125,10 @@ def test_send_success_without_id_with_additional_data(message):
     assert "ID and handler" in str(exc.value)
 
 
-def test_send_success_without_id_and_type(message):
+def test_send_success_without_id_and_type(consumer):
     with pytest.raises(Exception) as exc:
         send_success(
-            message=message,
+            consumer=consumer,
             message_id=None,
             handler=None,
             data={"some_field": "additional_info"},
@@ -133,65 +136,65 @@ def test_send_success_without_id_and_type(message):
     assert "ID and handler" in str(exc.value)
 
 
-def test_send_message(message):
-    send_message(message=message, message_id=1, handler="some_handler")
-    data = message.reply_channel.data
+def test_send_message(consumer):
+    send_message(consumer=consumer, message_id=1, handler="some_handler")
+    data = consumer.data
     assert len(data) == 1
     assert data[0]["id"] == 1
     assert data[0]["type"] == "some_handler"
 
 
-def test_send_message_with_additional_data(message):
+def test_send_message_with_additional_data(consumer):
     send_message(
-        message=message,
+        consumer=consumer,
         message_id=1,
         handler="some_handler",
         data={"some_field": "additional_info"},
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert data[0]["id"] == 1
     assert data[0]["type"] == "some_handler"
     assert data[0]["data"]["some_field"] == "additional_info"
 
 
-def test_send_message_without_id(message):
-    send_message(message=message, message_id=None, handler="some_handler")
-    data = message.reply_channel.data
+def test_send_message_without_id(consumer):
+    send_message(consumer=consumer, message_id=None, handler="some_handler")
+    data = consumer.data
     assert len(data) == 1
     assert "id" not in data[0]
     assert data[0]["type"] == "some_handler"
 
 
-def test_send_message_without_id_with_additional_data(message):
+def test_send_message_without_id_with_additional_data(consumer):
     send_message(
-        message=message,
+        consumer=consumer,
         message_id=None,
         handler="some_handler",
         data={"some_field": "additional_info"},
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert "id" not in data[0]
     assert data[0]["type"] == "some_handler"
     assert data[0]["data"]["some_field"] == "additional_info"
 
 
-def test_send_message_without_id_and_type(message):
+def test_send_message_without_id_and_type(consumer):
     send_message(
-        message=message,
+        consumer=consumer,
         message_id=None,
         handler=None,
         data={"some_field": "additional_info"},
     )
-    data = message.reply_channel.data
+    data = consumer.data
     assert len(data) == 1
     assert "id" not in data[0]
     assert "type" not in data[0]
     assert data[0]["data"]["some_field"] == "additional_info"
 
 
-def test_send_message_without_anything(message):
+def test_send_message_without_anything(consumer):
     with pytest.raises(Exception) as exc:
-        send_message(message=message, message_id=None, handler=None)
+        send_message(consumer=consumer, message_id=None, handler=None)
     assert "empty message" in str(exc.value)